+++ package.json
@@ -20,11 +20,13 @@
     "@nivo/pie": "^0.99.0",
     "@nivo/radar": "^0.99.0",
     "@tanstack/react-query": "^5.72.2",
+    "@types/react-router-dom": "^5.3.3",
     "axios": "^1.8.4",
     "html2canvas": "^1.4.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-joyride": "^2.9.3",
+    "react-router-dom": "^7.6.1",
     "react-toastify": "^11.0.5"
   },
   "devDependencies": {
